Removed modal confirm, added a very basic crack open door button
Changed open/close button to have no confirmation. It was too slow for my liking.

Added a new route to test out my very very basic double triggering of the relay 2 seconds apart to crack the garage. It does not do a check to see if the garage is open or closed before triggering.